requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
playwright==1.40.0
openai==1.3.5
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
pandas==2.1.3
pillow==10.1.0
PyPDF2==3.0.1
openpyxl==3.1.2
httpx==0.25.2
matplotlib==3.8.2
